+++ simple_history/utils.py
@@ -49,6 +49,7 @@ def bulk_create_with_history(
     objs,
     model,
     batch_size=None,
+    ignore_conflicts=False,
     default_user=None,
     default_change_reason=None,
     default_date=None,
@@ -82,8 +83,10 @@ def bulk_create_with_history(
 
     second_transaction_required = True
     with transaction.atomic(savepoint=False):
-        objs_with_id = model_manager.bulk_create(objs, batch_size=batch_size)
-        if objs_with_id and objs_with_id[0].pk:
+        objs_with_id = model_manager.bulk_create(
+            objs, batch_size=batch_size, ignore_conflicts=ignore_conflicts
+        )
+        if objs_with_id and objs_with_id[0].pk and not ignore_conflicts:
             second_transaction_required = False
             history_manager.bulk_history_create(
                 objs_with_id,